backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
celery==5.4.0
redis==5.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
greenlet==3.0.3
email-validator==2.2.0
gitpython==3.1.43
anthropic==0.52.0
pydantic==2.10.3
pydantic-settings==2.7.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
